+++ x/btcstaking/keeper/inclusion_proof_test.go
@@ -28,7 +28,7 @@ func FuzzVerifyInclusionProofAndGetHeight(f *testing.F) {
 		// mock BTC light client and BTC checkpoint modules
 		btclcKeeper := types.NewMockBTCLightClientKeeper(ctrl)
 		btccKeeper := types.NewMockBtcCheckpointKeeper(ctrl)
-		h := testutil.NewHelper(t, btclcKeeper, btccKeeper)
+		h := testutil.NewHelper(t, btclcKeeper, btccKeeper, nil)
 
 		// set all parameters
 		h.GenAndApplyParams(r)